handle error corner cases
When stripe throws an error, and the original form as an onclick event,
we briefly see the error before the form is submitted according to
the onclick event (i.e. the PreventDefault() and return false
does not have any effect).

This fix removes any onclick attributes at the beginning and only
re-adds them and resubmits the form if we should.