allow multiple stripe payment processor on back end
Without this patch, you can the error "no such token" if:

 * You have two different stripe payment processors configured.
 * you are making an offline contribution
 * using the provided payment processor back end, you choose the
 non-default payment processor.

The stripe javascript submits the payment using the default payment
processor, but the PHP codes tries to charge it using the user-selected
payment processor, which leads to the error.

This fix also ensures that if a non-stripe payment processor is chosen
we won't try to send it to stripe.