provide additional information about stripe events
At some point it would be nice to offer a web ui for users
to review recent events and try to "repair" any that did not
go through properly using the import apis. This is a pre-cursor
to that, which provides additional information on which parts
of the event did not get properly added to Civi.